requirements.txt
tensorflow-gpu==2.2.0
matplotlib==3.2